Refactor App to be a class component.
The rules behind when you are allowed to use certain hooks were
getting in my way and making it hard to write the App component
correctly. I'm not sure if there was a better way to go about
this, but every problem I was having with hooks could be solved
by simply creating a class component, so I took that route instead
of trying to kludge a way to get hooks to work.